Migration: store amounts in int64 instead of decimal. Stellar internally stores payment amounts as int64.

+++ src/bitbucket.org/atticlab/horizon/db2/schema/migrations/3_aggregate_expenses_for_accounts.sql
@@ -3,14 +3,14 @@
 CREATE TABLE account_statistics (
     address             character varying(64) NOT NULL,
     asset_code          character varying(12) NOT NULL,   
-    daily_income 		decimal NOT NULL DEFAULT 0,
-    daily_outcome 		decimal NOT NULL DEFAULT 0,
-    weekly_income 		decimal	NOT NULL DEFAULT 0,
-    weekly_outcome		decimal NOT NULL DEFAULT 0,
-    monthly_income		decimal NOT NULL DEFAULT 0,
-    monthly_outcome 	decimal NOT NULL DEFAULT 0,
-    annual_income		decimal NOT NULL DEFAULT 0,
-    annual_outcome		decimal NOT NULL DEFAULT 0,
+    daily_income 		bigint NOT NULL DEFAULT 0,
+    daily_outcome 		bigint NOT NULL DEFAULT 0,
+    weekly_income 		bigint	NOT NULL DEFAULT 0,
+    weekly_outcome		bigint NOT NULL DEFAULT 0,
+    monthly_income		bigint NOT NULL DEFAULT 0,
+    monthly_outcome 	bigint NOT NULL DEFAULT 0,
+    annual_income		bigint NOT NULL DEFAULT 0,
+    annual_outcome		bigint NOT NULL DEFAULT 0,
     updated_at 		    timestamp without time zone NOT NULL,
     PRIMARY KEY(address, asset_code)
 );
